+++ app.rb
@@ -13,6 +13,7 @@ require_relative './modules/create_person_3'
 require_relative './modules/create_book_4'
 require_relative './modules/create_rental_5'
 require_relative './modules/list_rentals_6'
+require_relative './modules/menu'
 
 class App
   # modules one for each option [1..6]
@@ -22,6 +23,7 @@ class App
   include CreateBook
   include CreateRental
   include ListRentals
+  include Menu
 
   def initialize
     @books = []
@@ -36,18 +38,6 @@ class App
     handle_command(command) unless command == 7
     puts 'Thank you for using this app!'
     exit
-  end
-
-  def display_menu
-    puts 'Please choose an option by entering a number'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts "7 - Exit\n "
-    print 'Number: '
   end
 
   def handle_command(command)
